+++ x/leverage/client/cli/query.go
@@ -42,7 +42,7 @@ func GetQueryCmd(queryRoute string) *cobra.Command {
 		GetCmdQueryMarketSize(),
 		GetCmdQueryTokenMarketSize(),
 		GetCmdQueryBorrowLimit(),
-		GetCmdQueryLiquidationLimit(),
+		GetCmdQueryLiquidationThreshold(),
 		GetCmdQueryLiquidationTargets(),
 		GetCmdQueryMarketSummary(),
 	)
@@ -628,13 +628,13 @@ func GetCmdQueryBorrowLimit() *cobra.Command {
 	return cmd
 }
 
-// GetCmdQueryLiquidationLimit returns a CLI command handler to query for the
-// liquidation limit of a specific borrower.
-func GetCmdQueryLiquidationLimit() *cobra.Command {
+// GetCmdQueryLiquidationThreshold returns a CLI command handler to query a
+// liquidation threshold of a specific borrower.
+func GetCmdQueryLiquidationThreshold() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "liquidation-limit [addr]",
+		Use:   "liquidation-threshold [addr]",
 		Args:  cobra.ExactArgs(1),
-		Short: "Query for the liquidation limit of a specified borrower",
+		Short: "Query a liquidation threshold of a specified borrower",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			clientCtx, err := client.GetClientQueryContext(cmd)
 			if err != nil {
@@ -643,11 +643,11 @@ func GetCmdQueryLiquidationLimit() *cobra.Command {
 
 			queryClient := types.NewQueryClient(clientCtx)
 
-			req := &types.QueryLiquidationLimitRequest{
+			req := &types.QueryLiquidationThresholdRequest{
 				Address: args[0],
 			}
 
-			resp, err := queryClient.LiquidationLimit(cmd.Context(), req)
+			resp, err := queryClient.LiquidationThreshold(cmd.Context(), req)
 			if err != nil {
 				return err
 			}